app/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.0
flask==3.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
python-dotenv==1.0.0
werkzeug==3.0.1
zipp==3.17.0
